requirements.txt
torch>=2.5.1
numpy>=2.1.3 
pyyaml>=6.0.2
